+++ defeatbeta_api/data/ticker.py
@@ -633,8 +633,8 @@ class Ticker:
         roa['report_date'] = pd.to_datetime(roa['report_date'])
 
         result_df = pd.merge_asof(
-            roe.sort_values('report_date'),
-            roa.sort_values('report_date'),
+            roe,
+            roa,
             left_on='report_date',
             right_on='report_date',
             direction='backward'
@@ -1302,6 +1302,31 @@ class Ticker:
         ).round(4)
         df.insert(1, "industry", industry)
         return df
+
+    def industry_equity_multiplier(self) -> pd.DataFrame:
+        roe = self.industry_roe()
+        roa = self.industry_roa()
+
+        roe['report_date'] = pd.to_datetime(roe['report_date'])
+        roa['report_date'] = pd.to_datetime(roa['report_date'])
+
+        result_df = pd.merge_asof(
+            roe,
+            roa,
+            left_on='report_date',
+            right_on='report_date',
+            direction='backward'
+        )
+
+        result_df['industry_equity_multiplier'] = round(result_df['industry_roe'] / result_df['industry_roa'], 2)
+
+        result_df = result_df[[
+            'report_date',
+            'industry_roe',
+            'industry_roa',
+            'industry_equity_multiplier'
+        ]]
+        return result_df
 
     def _quarterly_eps_yoy_growth(self, eps_column: str, current_alias: str, prev_alias: str) -> pd.DataFrame:
         url = self.huggingface_client.get_url_path(stock_tailing_eps)